requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.1            # via django
certifi==2020.11.8        # via requests
chardet==3.0.4            # via pypugjs, requests
django-environ==0.4.5     # via -r requirements.in
django==3.1.3             # via -r requirements.in
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via requests
oauthlib==3.1.0           # via requests-oauthlib
psycopg2==2.8.6           # via -r requirements.in
pypugjs==5.9.5            # via -r requirements.in
pytz==2020.4              # via django
requests-oauthlib==1.3.0  # via -r requirements.in
requests==2.25.0          # via requests-oauthlib
six==1.15.0               # via pypugjs
sqlparse==0.4.1           # via django
urllib3==1.26.2           # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
